requirements.txt
torch==2.0.1
torchvision==0.15.2
segmentation-models-pytorch==0.4.0
numpy==1.26.4
opencv-python==4.9.0.80                  
opencv-python-headless==4.10.0.84
albumentations==2.0.5
matplotlib==3.9.4 
pillow==10.4.0
onnx==1.14.0
pycocotools==2.0.7
PyYAML==6.0.1
scipy==1.13.0
onnxslim==0.1.31
gradio==4.31.5
opencv-python==4.9.0.80
psutil==5.9.8
py-cpuinfo==9.0.0
huggingface-hub==0.23.2
safetensors==0.4.4